vrend: do not set GL_TEXTURE_MAX_LEVEL to 0
In ANGLE, when GL_TEXTURE_MAX_LEVEL is changed, the underlying texture is
copied to a new texture and released. When we are using an imported texture
this breaks when the other context updates the texture. However, a max
level of 0 doesn't do anything meaningful anyways and is used as a ignored
value. That means we can avoid this issue in ANGLE by simplying not setting
GL_TEXTURE_MAX_LEVEL when it is 0.